# Conflicts:
#	libs/atlmfc/mfc.vcxproj
#	libs/atlmfc/mfc.vs2015.vcxproj
#	libs/rdestl/rdestl.vcxproj
#	libs/rdestl/rdestl.vs2015.vcxproj
#	src/NetBox/CMakeLists.txt
#	src/include/rtti.hpp

+++ src/core/SecureShell.h
@@ -84,7 +84,7 @@ private:
   UnicodeString FUserName;
   bool FUtfStrings;
   DWORD FLastSendBufferUpdate;
-  int FSendBuf;
+  intptr_t FSendBuf;
 
 public:
   static TCipher FuncToSsh1Cipher(const void * Cipher);
